services/payment_service/requirements.txt
Django>=5.2,<5.3
djangorestframework==3.17.1
psycopg2-binary==2.9.11
gunicorn==23.0.0
